Add incident-summary component
The naming with the nodes right now isn't super swell. This component
became summary but it's part of a section called summary.

But I'll mix that up together when I've done something about the form.
Because it feels like that's a good place for another component about
now, since then I can bundling the logic up nicer.

Then again, should just start with the next set of interactions:
yay/nay buttons for the action so it can become a past action.

When I have those together I can start making the proper summaries