Add descriptions of the tool
I wrote this because I was struggling with how I was going to the
distinction between "action" and "task" as I have described here.

I feel like there is a third generic term lurking here, or a more
specific, but I don't know what it is so this is enough.