Add single place to define events
Not loving this pattern but I can live with it until I can think more
about how the events should represent themselves and be emitted.

Overall, the dispatcher currently handles multiple things;

1. Dispatching
2. Creating events
3. Providing event data

The creation of events should probably be different, but also, the
event creator should be unique and not know about all events. Possibly,
that I should make unique event dispatchers per "grouping" of events
since there are relationships between them.

For example, there will be a grouped set of Action events, and similar
for Incident. So that's probably the next step for this.

And yeah, I'm catching myself consistently journaling out the next
steps in my commit messages and that's not a great pattern. But also,
whatever for now.