Add a task done button
To indicate the tasks are finished. Can tasks fail? I guess they can,
because I could fail to escalate and then continue. But is that more a
problem of maybe doing too narrow? But indicating changing escalations
can also be helpful… So maybe it's more whether this task is intended
to mitigate the incident?