Allow updating of active action
Not the prettiest code in the active action component but I'll loop
back to that later. I am liking how it works out with the event:

1. Dispatch event
2. Listen to the event
3. Update relevant attribute on component
4. Component acts on it

This way, when I move to multiplayer the event will continue to be
listened to. And I can do special actions on the first one, like I am
doing with the Countdown here, because I need to set the new expire time.

Closes #5.