Add EventListeners for listening to events
Relying on native HTML elements is nice for unintended places, but if
I'm anyway relying on knowing the event name and all that, then I might
just as well use the real object and work from there.

And this is a start to decouple the event creation and real sending of
events. Which will be more important as soon as I start adding storage.