Make updates referencing the parent event
The old way was that we sent the parent event's ID as the ID for all
subsequent updates, but it also mean that we'd send two IDs that were
for two completely different events.

I was wondering whether I should go with a `parentId` as a generic name
for all parents, but felt the added verbosity was worth it because less
risk of using it wrong. Let's see if that decision was a good idea
or not.